Create a separate file for image interface
